+++ platforms/native/components/Box/index.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components/native'
-import { ShallotProp, getStyle } from '@shallot-ui/core'
-import { withBoxStyleProps } from '@shallot-ui/box'
+import { getStyle } from '@shallot-ui/core'
+import { BoxShallot, withBoxStyleProps } from '@shallot-ui/box'
 
-const RawBox = styled.View<{ shallot?: ShallotProp }>(getStyle)
+const RawBox = styled.View<{ shallot?: BoxShallot }>(getStyle)
 export const Box = withBoxStyleProps(RawBox)