+++ components/Column/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shallot-ui/column
 
+## 0.6.0
+
+### Minor Changes
+
+- Removes forward ref to improve surface types
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/box@0.6.0
+  - @shallot-ui/core@6.2.0
+
 ## 0.5.0
 
 ### Minor Changes