+++ components/Modal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @shallot-ui/modal
 
+## 0.5.0
+
+### Minor Changes
+
+- Removes forward ref to improve surface types
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/text@0.7.0
+  - @shallot-ui/box@0.6.0
+  - @shallot-ui/row@0.6.0
+  - @shallot-ui/core@6.2.0
+
 ## 0.4.16
 
 ### Patch Changes