+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/utils
 
+## 4.1.0
+
+### Minor Changes
+
+- Removes forward ref to improve surface types
+
 ## 4.0.5
 
 ### Patch Changes