+++ platforms/web/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @shallot-ui/web
 
+## 0.45.0
+
+### Minor Changes
+
+- Removes forward ref to improve surface types
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/form-column@0.5.0
+  - @shallot-ui/checkbox@0.6.0
+  - @shallot-ui/form-row@0.5.0
+  - @shallot-ui/button@0.7.0
+  - @shallot-ui/column@0.6.0
+  - @shallot-ui/switch@0.6.0
+  - @shallot-ui/input@0.7.0
+  - @shallot-ui/fold@0.5.0
+  - @shallot-ui/text@0.7.0
+  - @shallot-ui/box@0.6.0
+  - @shallot-ui/row@0.6.0
+  - @shallot-ui/tag@0.6.0
+  - @shallot-ui/core@6.2.0
+
 ## 0.44.0
 
 ### Minor Changes